Refactored data access into driver api and cleaned up main()
Created an api for registering and requesting storage drivers.
Moved flags into a configuration struct to remove command line
dependencies from server.go and other parts of the app.
Removed the pipeline api.
Renamed fancy_request to message and moved it to a new namespace to
prevent a cyclic dependency issue during build.